requirements.txt
pandas==2.2.2
scikit-learn==1.4.2

apache-airflow[celery,postgres,redis]==2.9.0

pyyaml==6.0.1
python-dotenv==1.0.1

jupyter==1.0.0
ipykernel==6.29.4

# Google Drive / S3
# gdown==5.1.0
# boto3==1.34.82

